Write a C program that prints the size of various types on the computer it is compiled and run on.

+++ 0x00-hello_world/6-size.c
@@ -8,7 +8,7 @@ int main(void)
 {
 char c;
 int i;
-long long int d;
+int d;
 
 
 printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
